soccer_support_AI/requirements.txt
ultralytics>=8.3.0
opencv-python>=4.9.0.80
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.11.0
pyyaml>=6.0.1
tqdm>=4.66.0
